backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
openai==1.43.0
python-dotenv==1.0.1
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.7
pydantic>=2.7.4
sqlalchemy>=2.0.25
